refactor: add global styles and restructure files
- Implemented global styling for the App component to maintain a consistent theme
- Refactored the project directory by moving style-related files into a new 'styles' directory
- Adjusted import paths to align with the new directory structure for styles
