+++ src/styles/globalStyle.ts
@@ -1,0 +1,15 @@
+/**
+ * @author Marcin Koziel
+ * @date 01/19/2024
+ */
+
+import { StyleSheet } from "react-native";
+
+const globalStyle = StyleSheet.create({
+  container: {
+    flex: 1,
+    paddingTop: 50,
+  },
+});
+
+export default globalStyle;
